few updates for fun
explanation is mostly in the comments (feel free to delete those, they dont really belong in the application, i just didnt want to make a bunch of different commits). It should be more extensible now, and would be even more so with if you wanted to create a small sqlite3 database to store the posts. You should look into searching the top level comments to find their sentiment and see how that matches the post title's. 

I did all the coding through github's editor so there may be some syntax or indentation errors.